requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.1
openai==1.12.0
Flask-SQLAlchemy==3.1.1 